Create security configuration profiles.
